+++ HISTORY.md
@@ -1,3 +1,13 @@
+## 0.47.8 (11/1/2017)
+- add standalone sensor component (.addSensor)
+- add deviceState service, helps tracking and caching devices
+- add support to define device timing (power on, source switch and power off)
+- add support for optional initialise function of your controller
+- add support for device power state, send notification to NEEO Brain when device change its power state
+- update discovery docs, explain that discovery supports synchronous return values and promises
+- update notification service, prevent duplicate messages sent to brain
+- fixed typo "addButtonHander" instead "addButtonHandler", thanks @rhuss
+
 ## 0.45.7 (8/23/2017)
 - fixes error when unregistering the SDK from a Brain
 - fixes parameter order in device builder documentation